+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added grouping by `<optgroup>` tag for `options()` method in `Select` field.
 - `admin_auth()` helper.
 - Relation search via ajax for `Tags` field.
+- Allow to delete image if it's selected but not uploaded for `Image` field.
 
 ### Fixed
 - Fixed css for pagination block for `light` theme.